requirements.txt
fastapi
uvicorn
python-stopwatch
pyyaml